fix(frontend): Add jest ts dependencies
