+++ web/app/package.json
@@ -16,16 +16,18 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
+    "@types/jest": "^29.2.4",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "axios": "^0.27.2",
     "framer-motion": "^6",
+    "jest": "^29.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
+    "ts-jest": "^29.0.3",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
@@ -34,6 +36,11 @@
     "build": "react-scripts build",
     "test": "react-scripts test --reporters=default --coverage --testResultsProcessor jest-sonar-reporter",
     "eject": "react-scripts eject"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "!**/generated"
+    ]
   },
   "jestSonar": {
     "reportPath": "reports",
